feat: add notification handling, simple progress updates, uuu-rs refactor
